Auto-increment build number
Instead of manually managing our build number, we can instead rely on
the number of commits on master to determine the build number. This is
an easy and cheap way to keep this number consistently incrementing.

Note that we're _not_ automatically setting the
CFBundleShortVersionString. This is intentional. In previous versions of
this technique we set that to the last created tag, but that left us in
weird spots where we didn't want the tag created yet but needed to
create a tag _just_ so that this would work properly. I think it makes
sense for that to be a manually updated thing, while we automate the
build number.